Apply code review feedback
