+++ test/Microsoft.NET.Build.Tests/GivenThatWeWantToBuildADesktopLibrary.cs
@@ -130,7 +130,7 @@ public class NETFramework
 using System;
 public static class {project.Name}
 {{
-    {TestAsset.ConflictResolutionTestMethod}
+    {ConflictResolutionAssets.ConflictResolutionTestMethod}
 }}";
 
             var testAsset = _testAssetsManager.CreateTestProject(project)
@@ -145,7 +145,7 @@ public static class {project.Name}
                         new XAttribute("Include", "NETStandard.Library"),
                         new XAttribute("Version", "$(BundledNETStandardPackageVersion)")));
 
-                    foreach (var dependency in TestAsset.ConflictResolutionDependencies)
+                    foreach (var dependency in ConflictResolutionAssets.ConflictResolutionDependencies)
                     {
                         itemGroup.Add(new XElement(ns + "PackageReference",
                             new XAttribute("Include", dependency.Item1),
