+++ test/Microsoft.NET.Publish.Tests/GivenThatWeWantToPublishAHelloWorldProject.cs
@@ -227,7 +227,7 @@ public static class Program
         TestConflictResolution();
         Console.WriteLine(""" + outputMessage + @""");
     }
-" + TestAsset.ConflictResolutionTestMethod + @"
+" + ConflictResolutionAssets.ConflictResolutionTestMethod + @"
 }
 ";
             var testProjectInstance = _testAssetsManager.CreateTestProject(testProject, testProject.Name)
@@ -239,7 +239,7 @@ public static class Program
                     var itemGroup = new XElement(ns + "ItemGroup");
                     p.Root.Add(itemGroup);
 
-                    foreach (var dependency in TestAsset.ConflictResolutionDependencies)
+                    foreach (var dependency in ConflictResolutionAssets.ConflictResolutionDependencies)
                     {
                         itemGroup.Add(new XElement(ns + "PackageReference",
                             new XAttribute("Include", dependency.Item1),
